refactor(148): Unifica projetos sonarcloud
