--- .github/workflows/sonarcloud-create-auth-challenge.yml
@@ -1,37 +1,0 @@
-
-name: SonarCloud create-auth-challenge
-# name: "SonarCloud" + <nome da aplicação ou function Lambda>
-# Em monorepos, cada aplicação precisa ter seu próprio workflow de CI do SonarCloud, 
-# pois dessa forma o SonarCloud consegue analisar cada aplicação como projetos separados.
-# Mais informações em https://docs.sonarsource.com/sonarcloud/advanced-setup/monorepo-support/
-on:
-  push:
-    branches:
-      - main
-    paths:
-      - 'create-auth-challenge/**' # Diretório onde se encontra a aplicação dentro do monorepo
-  pull_request:
-    types: [opened, synchronize, reopened]
-    paths:
-      - 'create-auth-challenge/**' # Diretório onde se encontra a aplicação dentro do monorepo
-jobs:
-  sonarcloud:
-    name: SonarCloud
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0  # Shallow clones should be disabled for a better relevancy of analysis
-      - name: Install dependencies
-        working-directory: ./create-auth-challenge
-        run: npm ci
-      - name: Test and coverage
-        working-directory: ./create-auth-challenge
-        run: npm run test --coverage
-      - name: SonarCloud Scan
-        uses: SonarSource/sonarcloud-github-action@master
-        with:
-          projectBaseDir: create-auth-challenge # Diretório onde se encontra a aplicação dentro do monorepo
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}  # Needed to get PR information, if any
-          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
