--- create-auth-challenge/sonar-project.properties
@@ -1,19 +1,0 @@
-sonar.projectKey=rms-cognito-triggers_create-auth-challenge
-sonar.organization=grupo-g03-4soat-fiap
-
-# This is the name and version displayed in the SonarCloud UI.
-#sonar.projectName=rms-cognito-triggers_create-auth-challenge
-#sonar.projectVersion=1.0
-
-
-# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=.
-sonar.test=tests
-sonar.coverage.exclusions=tests/**/*.test.ts
-
-# Encoding of the source code. Default is default system encoding
-#sonar.sourceEncoding=UTF-8
-
-# O parâmetro abaixo foi adicionado manualmente pelo Danilo a fim de enviar o resultado do test coverage para o SonarCloud.
-# Mais informações em https://docs.sonarsource.com/sonarqube/9.8/analyzing-source-code/test-coverage/javascript-typescript-test-coverage/#analysis-parameter
-sonar.javascript.lcov.reportPaths=./coverage/lcov.info
