--- define-auth-challenge/.npmignore
@@ -1,1 +1,0 @@
-tests/*
