+++ events/define-auth-challenge/defineAuthChallengeTest.json
