+++ jest.config.ts
@@ -11,5 +11,5 @@ export default {
     collectCoverage: true,
     coverageDirectory: 'coverage',
     coverageProvider: 'v8',
-    testMatch: ['**/tests/unit/*.test.ts'],
+    testMatch: ['**/tests/unit/**/*.test.ts'],
 };
