--- pre-signup/.eslintrc.js
@@ -1,15 +1,0 @@
-module.exports = {
-    parser: "@typescript-eslint/parser",
-    parserOptions: {
-      ecmaVersion: 2020, // Allows for the parsing of modern ECMAScript features
-      sourceType: "module" 
-    },
-    extends: [
-      "plugin:@typescript-eslint/recommended", // recommended rules from the @typescript-eslint/eslint-plugin
-      "plugin:prettier/recommended" // Enables eslint-plugin-prettier and eslint-config-prettier. This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
-    ],
-    rules: {
-      // Place to specify ESLint rules. Can be used to overwrite rules specified from the extended configs
-      // e.g. "@typescript-eslint/explicit-function-return-type": "off",
-    }
-  };
