+++ sonar-project.properties
@@ -7,7 +7,13 @@ sonar.organization=grupo-g03-4soat-fiap
 
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-#sonar.sources=.
+sonar.sources=src/
+sonar.test=tests/
+sonar.coverage.exclusions=tests/**/*.test.ts
 
 # Encoding of the source code. Default is default system encoding
 #sonar.sourceEncoding=UTF-8
+
+# O parâmetro abaixo foi adicionado manualmente pelo Danilo a fim de enviar o resultado do test coverage para o SonarCloud.
+# Mais informações em https://docs.sonarsource.com/sonarqube/9.8/analyzing-source-code/test-coverage/javascript-typescript-test-coverage/#analysis-parameter
+sonar.javascript.lcov.reportPaths=./coverage/lcov.info
