+++ src/handlers/create-auth-challenge.ts
