+++ src/handlers/define-auth-challenge.ts
