+++ src/handlers/verify-auth-challenge.ts
