+++ template.yaml
@@ -14,8 +14,7 @@ Resources:
   PreSignUpFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: pre-signup/
-      Handler: app.lambdaHandler
+      Handler: pre-signup.lambdaHandler
       Runtime: nodejs20.x
       Architectures:
         - x86_64
@@ -26,12 +25,11 @@ Resources:
         Target: "es2020"
         Sourcemap: true
         EntryPoints: 
-        - app.ts
+        - src/handlers/pre-signup.ts
   DefineAuthChallengeFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: define-auth-challenge/
-      Handler: app.lambdaHandler
+      Handler: define-auth-challenge.lambdaHandler
       Runtime: nodejs20.x
       Architectures:
         - x86_64
@@ -42,12 +40,11 @@ Resources:
         Target: "es2020"
         Sourcemap: true
         EntryPoints: 
-        - app.ts
+        - src/handlers/define-auth-challenge.ts
   CreateAuthChallengeFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: create-auth-challenge/
-      Handler: app.lambdaHandler
+      Handler: create-auth-challenge.lambdaHandler
       Runtime: nodejs20.x
       Architectures:
         - x86_64
@@ -58,12 +55,11 @@ Resources:
         Target: "es2020"
         Sourcemap: true
         EntryPoints: 
-        - app.ts
+        - src/handlers/create-auth-challenge.ts
   VerifyAuthChallengeFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
-      CodeUri: verify-auth-challenge/
-      Handler: app.lambdaHandler
+      Handler: verify-auth-challenge.lambdaHandler
       Runtime: nodejs20.x
       Architectures:
         - x86_64
@@ -74,7 +70,7 @@ Resources:
         Target: "es2020"
         Sourcemap: true
         EntryPoints: 
-        - app.ts
+        - src/handlers/verify-auth-challenge.ts
 
 Outputs:
   # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
