--- verify-auth-challenge/.prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-    semi: true,
-    trailingComma: "all",
-    singleQuote: true,
-    printWidth: 120,
-    tabWidth: 4
-  };
